+++ lib/debug_me.rb
@@ -1,7 +1,14 @@
 require 'pp'
 require_relative 'debug_me/version'
 
-$DEBUG_ME = true
+# Initialize $DEBUG_ME from environment variable or default to true
+# Supports common boolean representations: true/false, yes/no, 1/0, on/off
+if ENV.key?('DEBUG_ME')
+  env_value = ENV['DEBUG_ME'].to_s.downcase
+  $DEBUG_ME = !['false', 'no', '0', 'off', ''].include?(env_value)
+else
+  $DEBUG_ME = true
+end
 
 DebugMeDefaultOptions = {
   tag:    'DEBUG',  # A tag to prepend to each output line